Added some django migrate and django check in makefile
It seems that the `manage.py migrate` command will be often used,
therefore add it as makefile target to simplify its uses.

Also added the `manage.py check` command to the lint target. This
command is kind of a django linting command as it does some static code
checking (see https://docs.djangoproject.com/en/3.1/ref/checks/). To
start with the WARNING level but this could change in the future as we
gain experience with django.