BGDIINF_SB-1344: Uses django test runner instead of nose
Nose has the advantage of being able to capture logs but don't allow to
configure the database initialization as does the django runner. To
avoid issues with testing using the DB I changed the runner to be the
one from django. I added a very simple configuration to disable logging
on the local development in order to have a clean test output. The
logging can still quickly be enabled using TEST_ENABLE_LOGGING
environment variable.