BGDIINF_SB-1343: Removed exception based on environment variable in django config
The django config is used by ./manage.py script and should not be
dependent on environment variable being set, therefore we should have
always default value. For the DB settings we use the same default value
as in .env.default.

Note this is necessary for building the production docker image which
calls the ./manage.py collectstatic and building container should not
depends on environment variable.