Fixed gunicorn dependency issue
A new version of the package greenlet which is a dependency of gevent
has been released. Because this version was not explicitely given in our
requirements and that our current gevent (used by gunicorn) specifyied
a version >=0.4.16, greenlet was automatically updated to 0.4.17.
Unfortunately this version is not compatible with gevent 20.6.2,
therefore I updated to 20.9.0 to avoid such issue. This should not be an
issue in future as we are now using pipenv.

Also update the PipFile version to not lock the versions of package in
order to allow update of minor version.