BGDIINF_SB-1266: Updated Collection, Item and Asset models
Changed the Link relationship to be many-to-one instead of many-to-many
this is hte correct relationship for the links.

Set the correct relationship between assets and item.

Removed the location field from the Item and Asset model as this one is
not needed. It can be generated when writting the object.

Also changed the variant type in Asset which should be a single string.