BGDIINF_SB-1362: Added Items and Item endpoint and corrected the Collections endpoint
Moved the collection view back to views.py as it was internal discussed.

In the collections endpoint encapsulate the list into an object in the
'collections' key as defined in the spec.

Fixed a bug in the collection endpoint where the collection name is a
string and not a slug. Slug doesn't allow special characters like `.`
which might be part of the collection name.

Added Items and Item endpoint.