BGDIINF_SB-1374: addapted unit test for pagination
Creates also severals objects in the DB in order to properly test
pagination. Unfortunately testing the 'next' url doesn't work in
unittest, but it has been manually tested and it worked.

Removed test docstrings because they are displayed in the test ouptut
with verbosity=2 and this make the output hard to read.