Clean up spec
Uses >- instead of |- for descriptions in yaml. This allow to have the
code in yml not longer than 100 characters which make the yaml file
easier to read. Another advantage is the rendered output has a dynamic
lines break. Wiht >- it requires a double line break to keep them in
output.

Also make sure that description and summary starts with an upper case.