BGDIINF_SB-1371: Added datetime query in items endpoint (fixed PR #68)
This fixed the wrong merge of PR #68 which forced pushed a WIP commit.

Also make sure that we use datetime object when creating directly
models. A model should always use python native data, that's the goal of
an ORM, however django doesn't convert datetimes string when
instanciating directly a model, but keeps it as string ! That's
apparently not the case when using serializer and/or forms and/or the
django admin page.