BGDIINF_SB-1434: Updated logging-utitilies with log extra and json formatter bug fix
The logging-utilities version 1.1.0 had a bug in the json formatter
where the log would crash if the add_always_extra was true and an extra
was not serializable into JSON.

The new version also supports extra formatting in the standard output.

Removed unecessary logs that are duplicate due to the logging
middleware.