BGDIINF_SB-1401: Fixed DEBUG=True prod image and renamed docker target test
The docker production image could not be started with the DEBUG flag set
to true. This was due to the fact that the DEBUG=True required some
django app that were only configured in the settings_dev.py and not in
prod settings.

Now in the prod image the DEBUG flag is forced to false to avoid issue
when the flag is externally set (e.g. deploying a production image on
dev staging). Anyway we should never deploy an image on production with
django DEBUG enabled.

Also renamed the docker image target test to debug as this target is not
olny used for testing in the CI but also used on the k8s dev staging.