BGDIINF_SB-1402: Fixed link to api doc
The links to the API documentation (html and yaml) had a double // in
front of `static` which broke the link.

To fix it I removed the trailing slash in API_BASE settings, because it
make the code more explicit and is less prone to errors as it happened
with the static links of the api doc.