BGDIINF_SB-1385: Corrected serializer validation
In a model serializer when explicitely declaring the fields, we need to
add the validation manually and these validations should then match the
one from the models.
Note that when the fields are not manually declared but simply listed in
the Meta.fields then the validation is automatically generated with the
validator from the models.