BGDIINF_SB-1500: Fixed collection delete stackoverflow
When deleting a collection containing items, django crashed with a
stackoverflow. This was due to the code that saved the original fields
value in the Item.__init__() method. Why this caused a stackoverflow it
is still a mistery to me.

But instead of using this hack I changed the algorithm to use a django
way of keeping track of original as described in https://docs.djangoproject.com/en/3.1/ref/models/instances/#customizing-model-loading
and this solved the issue.