BGDIINF_SB-1482: Added Item POST, PUT, PATCH, DELETE Endpoints
In order to handle the PUT of optional field correctly (when field is
missing in PUT request, then it needs to be deleted in DB, aqua set to
null) the model/serializer has been updated. For this the serializer
needs to have a default value set to None and should not allow null or
blank value.