BGDIINF_SB-1521: Added ETag support for conditial views
Added the ETag to the GET Answer header and add support for the If-Match
and If-Not-Match in all endpoints.

Also consolidate and harmonized the collection update functions. This
was needed in order to make the code easier to understand/read with the
addition of the etag update. This also improved the code by avoiding
saving a collection several times in a row or when not needed at all.
This also improve the code consistency.