create docker-compose dev setup
In order to simplify the setup for local development, the
necessary utils Postgres and MinIO are wrapped in a docker-compose file.

The existing docker-compose.yml used in CI is renamed to
docker-compose-ci.yml, the buildspec.yml was adapted to use that file.

Note: the version of mock must be locked to 4.0.2 since moto
has issues with versions 4.0.3, see
https://github.com/spulec/moto/issues/3535