BGDIINF_SB-1525 moved validation of JSON payload in validate() in serializers.py
Doing validation of JSON payload (checking for extra or read-only fields) in the
views.py by overriding get_serializer() worked but yielded a 500 instead of 400
when trying to post extra or read-only fields by using the browsable API views.

Hence the class CommonJsonValidatorSerializer was created in serializers.py
with an overriden validate() in which the JSON payload validation is done.
Additional validations for ItemSerializer were implemented, too.