BGDIINF_SB-1512: Fixed ./manage.py runserver
Running the server locally using django manage.py did not work anymore
due to the gevent monkey patching and wrong import order.

The import order has been corrected and the pragma `isort:skip_file`
will prevent future issue with isort on this file.

The import order fix did not solve the monkey patching crash, therefore
it is commented out for the moment. We will need to dig further to fix
it properly.