BGDIINF_SB-1512: Raise an error when configuration is missing
This is need to avoid having boto3 crashing with a name error which then
requires debugging to know what it is happening.

To avoid trouble with the CI with the settings raising errors, we use
the .env.default settings in dockerfile and also sets the configuration
for the ci docker file for testing.