BGDIINF_SB-1534: avoid duplicate validation in items
The item properties datetimes were validated twice in a row. This was
unecessary and due to a wrong uses of the django/rest_framework in
unittest. Creating an object with
ModelClass.objects.create(**validated_data) should
only be done with validated data, like for example in the serializer.
This method doesn't do any validation. If we don't have validated data,
then we need to use ModelClass(**unvalidated_data).full_clean().save()

So the item model unittest has been corrected in order to remove the
self.clean() in the item.save() method. This introduce a sligthly
performance improvement, but more important it makes the code cleaner.