BGDIINF_SB-1534: improved data factory for unittest
For asset and item in the json output the relation name of the item or
collection was part of the output. When testing post/put API the
relation should not be part of the payload but from the request path.

Now the data factory allow to tell if the json is to be used for
post/put/patch  request or for get/serialize methods. This made the
unittest more correct.