BGDIINF_SB-1576: Removed empty asset fields from GET request
The description, title, geoadmin_variant and geoadmin_lang were not
always stripped in the GET output when they were empty. That was the
case for the description when it was generated from the admin page (but
not when generated via POST). This was also the case for geoadmin_lang
when generated via Admin page but not from POST.