Fixed Item PATCH; missing 'type' in answer
When doing a PATCH on an existing Item, in the PATCH answer the
attribute 'type' was missing. Apparently the ReadOnlyField don't works
with the partial update. Therefore moved the 'type' attribute into a
SerializerMethodField, which fixed the issue.

The issue has been found with modification of the test base class where
in next commit we also check for auto generated and fix attributes of
the STAC API.