Issue BGDIINF SB 1548
Readme was supposed to be partially rewritten to reflect the fact that
the standard way of setting up the database for local development is
through docker containers.

I also rewrote a little bit about setting up the project, to have the
environment creation and project setup before the databases setup, as it
would most likely be accomplished in this order.