Do not update the collection extent and summaries on collection delete.
When deleting a collection from the admin page, the update extent and
summaries was triggered for each asset and items that was also deleted.

When deleting from the admin page, the model's delete() method are not
called, but the `pre_delete` signal is sent. With this signal we are
deleting asset on S3 using the File delete() method. This method
triggered and Asset update which was triggering a collection summaries
update and an item update which triggered a collection extent update.

Because we are deleting an asset (file and metadata) it doesn't make
sense to save the metadata object.

This improves massevely performance for deletion from admin page
(deletion of a collection with several thousand items did not work
before due to timesout)