BGDIINF_SB-1626: Fixed asset file validation cache issue
USing the HTTP HEAD request on the asset file on S3 can be dangerous on
the deployment because this request is suggest to caching. By default
the caching has an expiry time of 2 hours ! Which means that the answer
of the get HTTP HEAD might not be correct due to cache. This is quite
dangerous for validation.

Therefore now we use boto3 to check the asset file existance and to
check its checksum.

Added also a management command to upload dummy assets on S3 for testing
purpose.