BGDIINF_SB-1626: Fixed asset href issues
The Asset href attribute was missing in the PUT and PATCH response. This
was due to a bug in the hide_fields functionality of the serializer. The
href attribute was actually not needed in POST request and is making the
code complexer for nothing. Therefore this field has been changed to be
read only (as he was already defined in the spec) which allowed to
remove the buggy hide_field feature.

Another bug popout together with the href field as read only and the use
of S3 for asset file verification, was that the asset renaming would failed
due to trying to verify the asset file existance in the destination path
before the renaming. This is also now solved where we make sure to test
the source file location in case of renaming.