Improved logging.
Now the logging is redirected automatically to log files for the
management command, only management log with level WARNING and higher
plus all ERROR logs are still printed in the console, the rest is in log files

Also the default log config for running the django server locally, sends
logs in files in plus as the console. The django logs have then INFO
level on the console but DEBUG level on the files.