Totaly remove the TEST_ENABLE_LOGGING env var
This is not needed anymore as logging is done in a file for local
unittest and can be completely disabled with LOGGING_CFG=0. Note in
order to disable logging in unittest we need to use the override flag in
the load_dotenv() otherwise the value of the .env.local won't be used
but the default set in the manage.py is used instead.