ISSUE BGDIINF_SB_1631: wrong status on items and assets queries
Issue : When trying to access the items from a non existing collection,
or when trying to access the assets from one non existing item, the
application would return a 200 status with an empty list of items or
assets.

Fix : using the get_object_or_404() method allows consistent comportment
with other endpoints which returned a correct 404 not found status.