Issue Description :   It was asked to use the official confederation CSS in the API spec documentation.   Using the swiss confederation logo, as well as links to the geoportal and the legal department of swisstopo was also within the tasks.
Fix (Partial) : we imported only the 'build' directory from the official
swiss styling github : https://github.com/swiss/styleguide. We also
added a link segment which gives the location of the css. The location
of the favicon was also changed to use the one from the styleguide.

Problem : Since api.html and apitransactional.html are generated with
Redoc, classes don't match anything with the official css, and we might
as well have no css (I did not see any difference between using the css,
and not calling any css).

I have not yet placed the logo nor the links at the bottom.