BGDIINF_SB-1664: Fixes and improvement of the admin Asset page
The asset admin page for create/update an asset could not be displayed
when the DB contained a lot of Items (several hundred of thousand). This
was due to the fact that the view was trying to load all items for the
select box.

Now this issue has been solved by using an auto_complete_fields.

Other than that some other improvements have been done:

- Moved the AssetAdmin at the end of the file to have a consistent class
order
- Added collection and item in the list view
- Allow the search of wholeword using ""
- Organized the asset fields in fieldset
- Displayed the collection name in the asset view