BGDIINF_SB-1667: Fixed asset update/deletion crash
When deleting an Asset that had no eo:gsd or proj:epsg or
geoadmin:variant (value as None) in a collection with non empty
summaries, a `ValueError("list.remove: x not in list")`. Same for
updating an Asset without eo:gsd or proj:epsg or
geoadmin:variant (value as None) with a value.