BGDIINF_SB-1681: Log exception also on production build
On production build with the DEBUG=False the exception where not logged
because the rest framework custom exception handler in apps.py did
caught all exceptions and returned a proper 500 message (JSON or HTML
depending on Accept header or format query).

So now we logs the exception in the custom exception handler instead of
in the exception middleware. This means that for all exceptions on the
API endpoints we will have a log message from logger="stac_api.app" with
level CRITICAL and with exc_info=True. For all exceptions happening from
the admin interface endpoints, we will have a log from
logger="django.server" of level ERROR with exc_info=True