Updated pylint and fixed duplicate-code, raise-missing-from
In PR #249, django has been updated and it triggered a lots of no-member
pylint errors when referencing django models. This can be solved by
updating pylint to a version >= 2.6.

The update of 2.6 added a new raise-missing-from check which produced some
errors in our current code. As this check is very usefull, see also
https://stefan.sofa-rockers.org/2020/10/28/raise-from/ I fixed these
errors properly. Note I used `from None` for the ValidationError as
these ones are caught and translated into HTTP 400 errors, they don't
have any backtrace.

The current pylint version 2.7.0-2.7.2 have fixed an issue with the
similarities check, where the check was not done at all when runnings
pylint --jobs=2+, which we do. This showed several duplicate-code issue,
some of them where very relevant, like the pagination duplicate unittest
code and also the bbox conversion to geometry that I moved into a
functions. Unfortunately these pylint version still suffers from a bug
concerning the similarities, the config min-similarity-lines is ignored
! For our case the default value of 4 is too small and this cause some
false positive, testing pylint version 2.6.2 --jobs=1 showed that having
a min-similarity-lines=8 don't have false positive, therefore I change
it to this value which seems to me a convenable value.
We need to see then in the next 2.7 release if the bug is fixed or if we
need to either add the next 2.7 to the ignore list or totally disable
the duplicate-code checker.

I personally find this checker very good as it allow to improves code
and I would try to keep it, unfortunately, currently the only way to
have it working correctly would be to set the --job=1 and then the
lintings takes ~50s instead of ~15s...