BGDIINF_SB-1647 filter assets for their collection and/or item
added autocomplete search fields and dropdown list filters for filtering
assets for their collection and/or items.
Following behaviour is implemented:
If a collection is defined in the collection filter, only those items will appear
in the dropdown list of the item filter, that are in the currently selected collection.
ItemAdmin's get_search_results() was overriden with a few hacky lines for this.
Otherwise, all items would appear in the list, which does not make sense.
Note: The responsiveness after clicking on those filters could be better. Its a bit
slow when there are many items or collections. Probably will improve this soon.