BGDIINF_SB-1723: Implemented Collection upsert
We also make use of the DB atomic transaction to do proper DB rollback
in case of errors.

The get_object() of the views have been removed in order to use the
default implementation using the lookup_field.

Also made all endpoints atomic to assure proper db rollback in case of
error.