BGDIING_SB-1724 get instance for serializer when UPSERTing already existing objects
In case of UPSERTing already existing objects, the serializer needs to be passed the correct instance.
For this self.get_object() ineeds to be called inside the UpdateInsertModelMixin. In case self.get_object() fails
with a Http404, instance will be set to None and the object will be created. Otherwise the instance will be passed
to the serializer and the object will be updated.