BGDIINF_SB-1724 added check_existence() and adapted UpdateInsertModelMixin
The unit tests for race conditions sometimes passed and sometimes failed. Avoiding the get_object() and catching a 404
probably helps to fix this.
Hence a check_existence() was added to the collections, items and assets view and the UpdateInsertModelMixin was adapted
to use this. If the instance already exists, it is passed to the serializer together with the serializer_kwargs. If it
does not exist yet, the serializer is passed the data only.