BGDIINF_SB-1739: Moved and reused view validation as well as etag fix
The view validation code to check if parent exists in list views has
been moved in a separate module in order to be better reused. The
performance has been slightly improved due to the fact that we only do
one DB hit by checking the direct parent which can only exists if its
parent exists.

The Asset Etag was also not computed correctly if two asset in two
different items would have had the same name which is allowed. So now
the query to compute the asset etag takes also the parents into account.