BGDIINF_SB-1724 check parent objects | SlugRelatedFields | adaptions in data_factory.py
Current way the SlugRelatedFields collection and item in item's and
asset's serializer allow creation of asset's with a wrong collection in
the path. Also problems can occur, when two assets of the same name
exist in different parent items, or two items of the same name in two
different parent collections respectively. Hence a few changes were
implemented:

* adding parent collections and/or items to the serializer's validated data so tupdate_or_create() would notice
  that they don't exist, when the validated_data is passed as the defaults argument
* removed the serializers.SlugRelatedField() item in asset serializer
* made SlugRelatedField collection in item serializer read_only
* made use of the get_object_or_404() for checking the existence of parent collections or items.
* adapted data_factory.py to not include collection in ItemSample's get_json()'s data and not include item in AssetSample()'s get_json()'s data,
  in case of method = deserialize
* small according adaptions in test_serializer.py