BGDIINF_SB-1739: Updated readme and pylintrc
* Updated the README.md table of content.
* Removed unused variable in Makefile
* Allowed module name with numbers in pylintrc: e.g. s3_multipart_upload
before only migrations module where allowed to have number in there
names. However a valid name must always start with a letter (except for
migrations).
* Also improved a unittest error message
* Fixed pagination bug in case the links would only contain the
paginations links. In this case `links` would have been a list of list
of link instead of a list of link. This use case what currently not used
but will be with the asset uploads view.