BGDIINF_SB-1739: Fixed missing number_parts bug in create asset upload
When creating an asset upload without specifying the number_parts, no
presigned url where created and the number_parts was set in db per
default to 1.

Now the number_parts is required and no default value is set in DB. This
would allow in future to change the SPEC to allow a create without
number_parts and create the part separately with a new endpoint leaving
the number of parts open.