BGDIINF_SB-1746: Removed the Item renaming functionality
Also improved the Asset admin page. In this one the read only item field
still had the help text for the field search functionality which don't
make sense for read only field. To avoid this simply changed the logic,
adding extra read only fields that are only displayed in the change view
and the write item field is only display in the add view.

Also improved the logging in case off renaming.