Improved admin page for Asset Upload
Now all fields are read only and the "update" button is hidden. We
should not change anything from the admin page for the Asset Upload and
do only changes from the API. This is because the asset upload have a
lots of functionality in the API view.

Delete of completed/aborted asset upload is accepted.