BGDIINF_SB-1844: #bugfix empty description field
Issue : when creating a provider through the admin UI and leaving the
description field empty, a description field would be present in JSON
outputs.

Fix : As the problem was that the admin interface saved an empty string
("") rather than a None object, we transform empty string descriptions
into None object before saving.