BGDIINF_SB-1864: Order API collections/items/assets by name
The view using pagination, the ordering is done in the pagination class.
This one did not had the ability to easily configure different ordering
per view. This has been added via the get_ordering() method which checks
for `ordering` attribute in the view.

For the assets ordering within items, we needed to do the ordering in
the prefetch related and in the get_queryset() because assets don't uses
pagination.