BGDIINF_SB-1897: eo:gsd value 0.0 not added to collection summaries
The eo:gsd value 0.0 was not added to the collection summaries which did
a crash when trying to unset the eo:gsd or change its value. A similar
issue was also present for the proj:epsg with value 0 (also value of 0
for espg don't make sense, but if by mistake it would be set, then it
would be an issue).

The bug has been fixed. Then because having a value of 0.0 for eo:gsd
doesn't make sense, we prevent it by raising a 400.