BGDIINF_SB-1897: Don't crash in case of removing missing summaries
When updating the collection summaries and trying to remove a summary
from a list, if the summary (eo:gsd, geoadmin:variant or proj:epsg) is
missing from the list, then the update crash preventing the collection
summaries update.

This case should not happen, but if the collection is
corrupted for any reason like another small bug or data have been
manually altered, we cannot recover from it.

Now in this case we can recover from such situation, removing an non
existent summary is simply ignored as the result is what we expect
anyway.

Also correct a possible bug with the geoadmin:variant where checking
with `if variant:` is not true when variant == ''. In theory the variant
is never an empty string but if it would be we would end up again in a
corrupted summaries that we cannot correct.