Prevent false positive no-member pylint issues
The new pylint 2.9 version is doing a deeper class member analysis,
however not always right. In the case of the data_factory.py module, it
resolve all `.model` to Model instance instead of its correct instance
(e.g. Item, Asset or Collection) and therefore triggers a lot of
`no-member` issues.