Fixed a crash when adding/updating item/asset with duplicate name in different collection
In theorie having duplicate object name (item or asset) in different
parent (collection or item) is allowed. In practice this is not the
case, but it can happen especially for testing purpose.

Creating and/or updating item with duplicate name over the API failed
with a crash (500; stac_api.models.Item.MultipleObjectsReturned: get() returned more than one Item)

This was because the upsert lookup was only done using the name and not
the parent(s) name.