BGDIINF_SB-2520: Avoid adding `{"duration": "duration"}` in logs without duration
When the `duration` attributes was not available in log record (in extra) then
the logs added a `"duration": "duration"` in json output. This was due to the
fact that the formatter could determine if the `duration` is a log record attribute
or a constant.

To avoid this we encapsulate the duration in the response object, this way it
is omitted if empty when the logging is configured so.