BGDIINF_SB-2799: Fixed performance issue with the search endpoint due to update_interval
In the search endpoint an aggregation over the whole queryset was done for
the update interval. This can potentially be very slow as a GET /search return
a queryset with all items in the DB ! Also the aggregation was done even
in a POST /search where we don't need to add Cache-Control header !

Now the cache-control header is only in GET /search geadded and the
update_interval aggregation is also only done for GET request.

The aggregation is done on the paginated queryset, meaning up to maximum of
100 items. We need to set the cache control header for the response of the
request and not based on all matching items.