BGDIINF_SB-2813: Fixed collection pagination due to missing title
Because the collections are ordered by title (was required for the STAC Browser)
the title is used by the cursor paginator. If the title is missing (null or empty
string) it breaks the pagination !

Therefore made the title mandatory. Up to now all production collection had a
title. This commit will add a default title as 'unknown' to all collection
without title, but this has no impact on productive data as they all do have
a title.