BGDIINF_SB-2827: Removed auto abort asset upload on duplicate upload
Only one asset upload is allowed at a time, before it used to automatically
abort ongoing upload when creating a new upload. The implementation had an
issue when multiple create upload happened on parallel, it crashed the service
due to race condition.

Making parallel or multiple create upload doesn't make sense as only one call
will be valid and the user will not know which one. E.g. if a user do two
requests in a row, both would have return 201 but only one of it would have
been valid and we could not ensure which one, could be the first or the second !

To avoid such random cases, we removed the auto abort mechanism. This means
that if multiple requests are sent at a time it is guarantied that only one
would be successful with 201 and valid. The others will returns 400 Bad Request
- Asset upload already in progress.

Improved a bit the test code.