BGDIINF_SB-2813: Fixed collection pagination
The collection paginations was broken because it was sorted by title and that
title could be missing or duplicates. For the cursor pagination it is a requirement
that the list is ordered by a unique existing column.

A first PR did fixed the issue by making the collection title required and unique
see #374. But because we have on PROD duplicate titles and that we need to keep
this functionality we revert the sorting by collection ID. See ticket for
more informations.