BGDIINF_SB-2158: add local prefix in docker build
Issue : As Lifecycle policies in our docker repositories allow, for the
moment, only prefixes to filter out images, local images without a
prefix have to be pruned manually, which is not efficient.

Fix : When building local images, a `local-` prefix is added in front of
the tag, which will enforce automatically self-cleaning policies for
local images pushed to our docker repository.