PB-409: fix broken health check and collections endpoint
Since this morning, the collections and health check endpoint are returning 500.
This is most likely caused by a special character in a newly updated collection description or ID field.
The error is, decode cannot decode a byte 0xc3 in position 199: unexpected end of data.
To prevent this, we now first decode the whole response.content and strip it afterwards, rather than the other way round.