PB-676: clean up specs Makefile a bit.
Add a %.props recipe for easier debugging of the spec building.

Remove the .PHONYness of the OpenAPI targets. These are fine regular build targets and this lets us say things like "make spec/static/spec/v1/openapi.yaml".

Add a "clean" target to allow easy force-rebuilding of the OpenAPI targets. Use that for the .PHONY "build-specs".