PB-676: remove duplicate YAML key from spec.
As far as YAML is concerned, each key in a Map must be unique. The "example" Map has two keys named "title". This change removes one of the two duplicate keys.

An alternative fix would involve turning this into a list of examples. However the OpenAPI validator we use does not seem to support it at this time.