PB-687: switch to a supported OpenAPI validator.
We currently use IBM's OpenAPI validator 0.54.0 (released circa February 2022)
from an old unsupported Docker image. With this change we update to version
1.19.2 (released June 2024) based on a Docker image shipped by IBM itself.

There were many changes in the validator since. In particular, the ruleset has
stricter checks about naming conventions that are specific to IBM guidelines
and which we don't follow. Rather than updating our specs to comply with IBM
guidelines, we disable these specific checks.

We also remove the old .validaterc configuration file which the newer validator
is not able to parse anyway.