PB-754: Refactor summary calculations
With many assets per collection, calculation of summaries is slow. As they are
calculated by triggers on the asset table, updating the asset table is also
slow.

Introduce counter tables to calculate collection summaries. This greatly
improves performance as the number of distinct values is small. In turn the
database size is slightly increased by adding more tables and data.