PB-694: make flaky serializer tests fail consistently.
The serializer tests depend on test running "fast enough" for timestamp
comparison to work. This is unreliable and causes inconsistent test failure.
This change adds a time.sleep(5) right after we record the start time of the
test such that timestamp comparison always fails.

This ensures the next change that attemps to fix the test actually fixes it
correctly. I plan to remove the delay afterwards.