PB-962: Improve performance to list items
Endpoint to get list of items is sometimes slow. This is due to the sql query
that the postgresql planner does not optimize properly, even though the indexes
seem to be set as expected. Using a subquery to filter greatly improves the
performance (up to 1 second reduction in query execution time).